pyproject.toml
[project]
name = "vvclaude-hooks"
version = "0.1.0"
description = "Claude Code hooks for Discord notifications and bidirectional communication"
dependencies = [
    "requests>=2.31.0",
    "discord.py>=2.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]